feat: add web-ui and wifi config

+++ PowerSample.h
@@ -7,5 +7,5 @@ struct PowerSample {
   float power_w;      // output power (after cycle constant)
   uint16_t crank_revs;
   uint16_t crank_evt_1024;
-  uint16_t adc_raw;   // Raw ADC value for calibration
+  float adc_raw;      // Raw ADC value (smoothed) for calibration
 };
